rename BRainNetworksInPython to scona
