--- BrainNetworksInPython/__init__.py
@@ -1,42 +1,0 @@
-"""
-BrainNetworksInPython
-========
-BrainNetworksInPython is a Python package for the analysis
-of structural covariance brain networks.
-Website (including documentation)::
-    http://whitakerlab.github.io
-Source::
-    https://github.com/WhitakerLab/BrainNetworksInPython
-Bug reports::
-    https://github.com/WhitakerLab/BrainNetworksInPython/issues
-Simple example
---------------
-FILL
-Bugs
-----
-Please report any bugs that you find `here <https://github.com/WhitakerLab/BrainNetworksInPython/issues>`_.
-Or, even better, fork the repository on GitHub and create a pull request (PR).
-License
--------
-FILL
-"""
-
-# Release data
-__author__ = "FILL"
-__license__ = "FILL"
-
-__date__ = "FILL"
-__version__ = 0.1
-
-__bibtex__ = """ FILL
-"""
-
-# imports listed in chronological order
-
-from BrainNetworksInPython.make_corr_matrices import *
-from BrainNetworksInPython.make_graphs import *
-from BrainNetworksInPython.graph_measures import *
-from BrainNetworksInPython.classes import *
-
-import BrainNetworksInPython.wrappers
-from BrainNetworksInPython import *
