--- BrainNetworksInPython/wrappers/__init__.py
@@ -1,2 +1,0 @@
-from BrainNetworksInPython.wrappers.corrmat_from_regionalmeasures import *
-from BrainNetworksInPython.wrappers.network_analysis_from_corrmat import *
