+++ docs/Makefile
@@ -91,9 +91,9 @@ qthelp:
 	@echo
 	@echo "Build finished; now you can run "qcollectiongenerator" with the" \
 	      ".qhcp project file in $(BUILDDIR)/qthelp, like this:"
-	@echo "# qcollectiongenerator $(BUILDDIR)/qthelp/BrainNetworksInPython.qhcp"
+	@echo "# qcollectiongenerator $(BUILDDIR)/qthelp/scona.qhcp"
 	@echo "To view the help file:"
-	@echo "# assistant -collectionFile $(BUILDDIR)/qthelp/BrainNetworksInPython.qhc"
+	@echo "# assistant -collectionFile $(BUILDDIR)/qthelp/scona.qhc"
 
 .PHONY: applehelp
 applehelp:
@@ -110,8 +110,8 @@ devhelp:
 	@echo
 	@echo "Build finished."
 	@echo "To view the help file:"
-	@echo "# mkdir -p $$HOME/.local/share/devhelp/BrainNetworksInPython"
-	@echo "# ln -s $(BUILDDIR)/devhelp $$HOME/.local/share/devhelp/BrainNetworksInPython"
+	@echo "# mkdir -p $$HOME/.local/share/devhelp/scona"
+	@echo "# ln -s $(BUILDDIR)/devhelp $$HOME/.local/share/devhelp/scona"
 	@echo "# devhelp"
 
 .PHONY: epub
