+++ docs/source/scona.datasets.NSPN_WhitakerVertes_PNAS2016.rst
@@ -1,10 +1,10 @@
-BrainNetworksInPython\.datasets\.NSPN\_WhitakerVertes\_PNAS2016 package
+scona\.datasets\.NSPN\_WhitakerVertes\_PNAS2016 package
 =======================================================================
 
 Module contents
 ---------------
 
-.. automodule:: BrainNetworksInPython.datasets.NSPN_WhitakerVertes_PNAS2016
+.. automodule:: scona.datasets.NSPN_WhitakerVertes_PNAS2016
     :members:
     :undoc-members:
     :show-inheritance:
