+++ docs/source/scona.wrappers.rst
@@ -1,21 +1,21 @@
-BrainNetworksInPython.wrappers package
+scona.wrappers package
 ======================================
 
 Submodules
 ----------
 
-BrainNetworksInPython.wrappers.corrmat_from_regionalmeasures module
+scona.wrappers.corrmat_from_regionalmeasures module
 -------------------------------------------------------------------
 
-.. automodule:: BrainNetworksInPython.wrappers.corrmat_from_regionalmeasures
+.. automodule:: scona.wrappers.corrmat_from_regionalmeasures
     :members:
     :undoc-members:
     :show-inheritance:
 
-BrainNetworksInPython.wrappers.network_analysis_from_corrmat module
+scona.wrappers.network_analysis_from_corrmat module
 -------------------------------------------------------------------
 
-.. automodule:: BrainNetworksInPython.wrappers.network_analysis_from_corrmat
+.. automodule:: scona.wrappers.network_analysis_from_corrmat
     :members:
     :undoc-members:
     :show-inheritance:
@@ -24,7 +24,7 @@ BrainNetworksInPython.wrappers.network_analysis_from_corrmat module
 Module contents
 ---------------
 
-.. automodule:: BrainNetworksInPython.wrappers
+.. automodule:: scona.wrappers
     :members:
     :undoc-members:
     :show-inheritance:
