+++ scona/__init__.py
@@ -1,0 +1,42 @@
+"""
+scona
+========
+scona is a Python package for the analysis
+of structural covariance brain networks.
+Website (including documentation)::
+    http://whitakerlab.github.io
+Source::
+    https://github.com/WhitakerLab/scona
+Bug reports::
+    https://github.com/WhitakerLab/scona/issues
+Simple example
+--------------
+FILL
+Bugs
+----
+Please report any bugs that you find `here <https://github.com/WhitakerLab/scona/issues>`_.
+Or, even better, fork the repository on GitHub and create a pull request (PR).
+License
+-------
+FILL
+"""
+
+# Release data
+__author__ = "FILL"
+__license__ = "FILL"
+
+__date__ = "FILL"
+__version__ = 0.1
+
+__bibtex__ = """ FILL
+"""
+
+# imports listed in chronological order
+
+from scona.make_corr_matrices import *
+from scona.make_graphs import *
+from scona.graph_measures import *
+from scona.classes import *
+
+import scona.wrappers
+from scona import *
