+++ scona/datasets/NSPN_WhitakerVertes_PNAS2016/500.centroids.txt
