+++ scona/scripts/annotate_images_to_movie.py
