+++ scona/scripts/fill_measure_dict.py
