+++ scona/scripts/make_supplemental_figures.py
