+++ scona/scripts/pysurfer_plot_500parcellation_surface_values.py
