+++ scona/scripts/read_fsaverage_info.py
