+++ scona/scripts/report_stats_tables.py
