+++ scona/scripts/useful_functions.py
