+++ scona/scripts/visualisation_commands.py
