+++ scona/wrappers/__init__.py
@@ -1,0 +1,2 @@
+from scona.wrappers.corrmat_from_regionalmeasures import *
+from scona.wrappers.network_analysis_from_corrmat import *
